[FIX] hr_payroll: fix error when printing employee payslip without contract
An error occurs while printing a payslip for an employee who does not have
a contract.

Steps to reproduce the error:
- Install the "hr_payroll" module
- Create an employee without contract
- Create a payslip for that employee > Compute Sheet > Print

Traceback:
Error while rendering the template:
TypeError: '<' not supported between instances of 'datetime.date' and 'bool'

https://github.com/odoo/enterprise/blob/368a81a277b2fa7c5eac7c63addd0d75ec89111b/hr_payroll/views/report_payslip_templates.xml#L70
Here, "version_id.contract_date_start" will be False when the employee does not
have a contract, which leads to the above traceback.

solution:
User can not print payslip without running contract.

sentry-6922017511

closes odoo/enterprise#96374

Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>