[FIX] stock_barcode: do not allow extra products from package
Issue 1: "Allow Extra Products" option ignored for packages

Steps to reproduce:

- In the settings enable "Packages"
- Go to Inventory > Configuration > Warehouse Management > Operation Types
- Disable "Allow Extra Products" on the "Delivery" operation type
- Create two storable product P1, P2 and add on hand quantities
    - 10 x P1 in a package PACK01
    - 10 x P2 in a package PACK02
- Create and confirm a delivery for 10 unit of P1
- Open your delivery from the barcode app
- Scan PACK02
> The content of PACK02 is added to the delivery even thought it
contains extra products.

Cause of the issue:

The check for extra products is only applied when scanning individual
products but is bypassed by package scan. To be more precise, the
`barcode_allow_extra_product` option is checked in the public method
`createNewLine`:
https://github.com/odoo/enterprise/blob/331442fd9cbd59d542432b2237299df9497ce241/stock_barcode/static/src/models/barcode_picking_model.js#L59-L80
While this method is called at new line creation when a product is
scanned, scanning a package will add new lines during the
`_processPackage` adn bypasses the rest of the `_processBarcode`:
https://github.com/odoo/enterprise/blob/331442fd9cbd59d542432b2237299df9497ce241/stock_barcode/static/src/models/barcode_model.js#L1261-L1267
The issue being that the `__processPackage` does not check the
`barcode_allow_extra_product` option and creates its new lines via the
private `_createNewLine` call:
https://github.com/odoo/enterprise/blob/331442fd9cbd59d542432b2237299df9497ce241/stock_barcode/static/src/models/barcode_picking_model.js#L1564-L1565
https://github.com/odoo/enterprise/blob/331442fd9cbd59d542432b2237299df9497ce241/stock_barcode/static/src/models/barcode_picking_model.js#L1655-L1667
https://github.com/odoo/enterprise/blob/331442fd9cbd59d542432b2237299df9497ce241/stock_barcode/static/src/models/barcode_picking_model.js#L1671

Fix:

Since scanning a package is expected to add all its content to the
 picking, and since a package can not be split among two locations, it
 is necessary to check in advance if any product of its content is extra
 and avoid any update in this case.

Issue 2: impossibility of package line removal

-> resolved since 19.0 see
https://github.com/odoo/enterprise/commit/060f4c3e24e7add05273cfad5853884546937d0e#diff-3c84b414a086808e446ae3f90b3b803edf964c28237c153af19edda3ed05cc6bL21

State of the art:

There is currently no option to remove a package line from the barcode.
In particular, once the option `show_entire_packs`(Move Entire Packages)
is enabled on a picking type, you can not remove the package line once
generated by a scan.

Steps to reproduce:

- In the settings enable "Packages"
- Go to Inventory > Configuration > Warehoue Management > Operation Types
- Enable "Move Entire Packages" on the "Delivery" operation type
- Create a storable product and add on hand quanties:
    - 10 units in package PACK01
    - 10 units in package PACK02
- Create and confirm a delivery for PACK01 (in the package lines)
- Open your delivery from the barcode app
- Scan PACK02
> The new line associated to PACK02 can not be removed by any mean

opw-4863621
opw-5080637

closes odoo/enterprise#97700

Forward-port-of: odoo/enterprise#97497
Forward-port-of: odoo/enterprise#96299
Signed-off-by: Steve Van Essche <svs@odoo.com>
Signed-off-by: Lancelot Semal (lase) <lase@odoo.com>