[FIX] account_{external_tax, avatax}: prevent error in when computing tax
Currently, an error occurs when computing taxes for an invoice.

Error:
    KeyError: 'manual_tax_amounts'

Cause:
When a user clicks the Compute Tax button, the system calculates the
tax for that invoice, including the manual_tax_amount [1]. However,
if no tax is found for a given customer ZIP code, the tax_values_list
becomes empty [2]. As a result, manual_tax_amounts in the base line
also becomes empty.

Later, when the system tries to fetch the manual_tax_amounts key from
extra_tax_data, it raises a KeyError [2], because the key no longer exists.

Additionally, the condition in [3] indicates that there is no guarantee
that the manual_tax_amounts key will always be present in the base_line.

FIX:
This commit ensures that if the manual_tax_amounts key does not exist,
an empty dictionary {} is used instead. This prevents the KeyError from
occurring during tax computation.

[1]: https://github.com/odoo/enterprise/blob/b17b6b4e5ca3085d831fc763457496b5c5b639c5/account_external_tax/models/account_external_tax_mixin.py#L155-L163

[2]: https://github.com/odoo/enterprise/blob/a51aee8f6e8bce2aa699d3199d6723495464a762/account_external_tax/models/account_external_tax_mixin.py#L88

[3]: https://github.com/odoo/odoo/blob/241c170dbece8c1652db9ca1aa935b2106ede532/addons/account/models/account_tax.py#L1330

sentry-6919182805

closes odoo/enterprise#96676

X-original-commit: a18c9c5c6b259b487d531c653bee630015989ccd
Signed-off-by: Joren Van Onder (jov) <jov@odoo.com>
Signed-off-by: Ashutosh Sharma (assh) <assh@odoo.com>