[FIX] timesheet_grid: correctly update the timesheet grid timer button
Before this commit, when the user changes the project from a timer
header inside the grid view of My Timesheets when a timer is running,
sometimes two grid rows show a timer is running at the same time. The
problem is because the timerRunning attribute of the grid row is not
correctly altered when the view is reloaded and so, it could lead to an
inconsistency.

This commit makes sure the GridTimerButtonCell only takes into account
the props it receives by the view instead of also check if the row is
timerRunning equals to true inside the model since the both should
return the same row.

Part-of: odoo/enterprise#96411
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>