[FIX] industry_fsm: correctly log the timer stopped on fsm task
Before this commit, when the user stops the timer in timesheets app, the
timesheet linked to the timer could be merged into previous timesheet,
the problem is we continue to read the timesheet afterwards to get the
task to log a note and could lead to a missing record error since the
timesheet is merged into another timesheet.

This commit makes sure the log message is correctly logged on the fsm
task without any issue.

Steps to reproduce:
==================
1. install industry_fsm module
2. Create a timer inside a fsm task
3. Stop it to create a new timesheet without editing the description
   field of the timesheet.
4. Start another timer inside the same task and stop the timer in the
   grid view of My Timesheets.

Expected Behavior:
-----------------
The timer should be stopped without any issue.

Actual Behavior:
---------------
MissingRecord error is raised because we try to read the timesheet after
having merged the timesheet into the previous one to get the fsm task
and post a message.

task-5074680

closes odoo/enterprise#96411

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>