[FIX] account_accountant: wrong context ID
On the bank reconciliation widget, you can click on a statement
amount to open the general ledger for the current journal. The action
to open it is called from JS, using the `active_id` value in the
context. However, when opening the bank reconciliation widget from an
invoice, the `active_id` is the invoice id and not the journal id, which
leads to an error when trying to open the general ledger.

Instead of `active_id`, we should use `default_journal_id`.

task-5108118

closes odoo/enterprise#96670

X-original-commit: a14476777b339cac3b4ad8791d096f1d14730257
Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>
Signed-off-by: Igor Bertrand (igbe) <igbe@odoo.com>