[FIX] *_spreadsheet_*: fix insertion of pivot sorted by date
If on the web pivot view we create a pivot with a date/dateTime field
in the columns, and sort on one of these columns, the resulting pivot
when inserted in a spreadsheet isn't sorted.

This was because we didn't normalize the date values in the pivot
`sortedColumn`, and we had UTC timestamps `2023-01-01 00:00:00` instead
of normalized values `01/2023`.

closes odoo/enterprise#99257

Task: 3575465
X-original-commit: 4693b9229da9ac4233fe2d96894e11eae07a8c7e
Signed-off-by: Adrien Minne (adrm) <adrm@odoo.com>
Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>