[FIX] stock_barcode: test_scrap_change_source_location
Sometime, the test `test_scrap_change_source_location` could fail
randomly. The issue happens in last few steps of the tour.
What we do is:
- We edit a move line lot in the form view;
- We save it;
- We validate the operation.

The validation is done by a barcode scan (`OBTVALI`) but since [1],
barcodes scanned when the user is somewhere else than in the barcode
lines view are skipped.
With a little bit of bad luck, the tour scans `OBTVALI` while the save
from the form view is not finished yet and thus, the scanned barcode is
ignored.

To reproduce that, run the test `test_scrap_change_source_location`
locally in debug mode and add a throttling (eg.: Fast 4G) before to run
the tour.

To solve the issue, finetune the `validateBarcodeOperation` default
trigger, so the error won't happen in this tour and other similar
contexts.

[1]: b3a855a870d1861515abaff8683081a39f95558f

Runbot build error: 232331

closes odoo/enterprise#96803

X-original-commit: 6cf23804d0b08ec9fb510fef9ddab3b8bd32b3ab
Signed-off-by: William Henrotin (whe) <whe@odoo.com>
Signed-off-by: Steve Van Essche <svs@odoo.com>