[FIX] iot: fix ctypes bytestrings
This PR fixes the bytestrings being sent as such when using ctypes C/C++
libraries in iot drivers.
After the PR https://github.com/odoo/odoo/pull/206903 the bytestrings
are not supported anymore in the requests sent to the database from the
iot and need to be decoded first.

opw-5129596

closes odoo/enterprise#96537

Signed-off-by: Louis Travaux (lotr) <lotr@odoo.com>