[FIX] iap_extract: fix loss of focus when clicking on image viewer
When the attachment showed by the viewer is an image and the selected
field is a date, a click on the image (to start a rectangular selection
or to click a box) would cause a loss of focus on the field which caused
the boxes to disappear, making it unusable.

This happened because the date fields are displaying a calendar popup
which listens on the `pointerdown` event to hide itself when the click
occurs outside of it. This causes the loss of focus of the field which
hides the boxes.

On top of this, we also need to prevent the propagation of the click
event when it occurs on the box layers, as it also causes the loss of
focus on x2many fields (there is a global click event listener that
exits the edit mode when the click occurs on another element).

These issues weren't noticed before as they only occur on image
attachments. On PDFs, the viewer is embedded in an iframe that prevents
the `pointerdown` and `click` events to be visible to the main document.

task-none

Part-of: odoo/enterprise#96975
Signed-off-by: Ricardo Gomes Rodrigues (rigr) <rigr@odoo.com>