[FIX] helpdesk: avoid traceback when closing ticket if no calendar
The issue: When you go to a helpdesk's team settings -> SLA Policies -> Working hours, set the working hours to empty and then disable SLA Policies and save.
After that if you try to move a ticket in the same team to done or canceled you will receive an exception.

Cause: The part of the code causing the issue is supposed to only run if a Working Hours policy is set.

Fix: Changed the section of the code to only run when Working Hours is set.

opw-5120962

closes odoo/enterprise#99948

X-original-commit: 58d55868750b827a9d5ebd8b4ab2cc23c4445eca
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Mohamed Jemai (mojem) <mojem@odoo.com>