[FIX] l10n_mx_edi: Fix certificate issue in cancellation
Steps to reproduce:
1. Install `l10n_mx_edi` with demo data
2. Switch to the ESCUELA KEMPER URGATE demo company
3. Create an invoice to the INMOBILARIA CVA demo partner
4. Send the invoice CFDI
5. Request cancellation of the CFDI
6. The cancellation fails with the error 'invalid passphrase'.

Analysis:
When calling `_finkok_cancel`, `_solfact_cancel`, or `_sw_cancel`,
one of the API call parameters is the `pem_key` of the certificate given
by the SAT.

Before 19.0, the PEM key was in an unencrypted format.

Since f88f8258ead, `env['certificate.key'].pem_key` is encrypted.

According to Finkok's API documentation, the private key should be
encrypted using DES when given as a SOAP parameter.
https://wiki.finkok.com/home/webservices/ws_cancelacion/cancel

However, in testing, encrypting the private key using DES seems to be
rejected by Finkok.

Solucion Factible and SwSapien don't indicate in their documentation
whether and how the private key should be encrypted.

Solution:
We send the private key unencrypted, as was already the case before
19.0.

opw-5137549

closes odoo/enterprise#96562

Signed-off-by: Laurent Smet (las) <las@odoo.com>