[FIX] hr_payroll: convert sql query to _read_group
Before this commit, when a compute needed to fetch the payslip lines
values, we need to flush all to be able to correctly get the values
thanks to a sql query. The problem is the `flush_all` will invalidate
all recordset in cache and will do more than expected.

This commit converts the sql query made to fetch the payslip lines
values into a read_group to be able to remove the flush_all and let the
orm invalidates the recordset/fields needed to correctly get what we
want.

closes odoo/enterprise#96741

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>