[FIX] l10n_be_reports: set partner accounts and new recon model
This commit sets AR and AP accounts of "VAT Partner" and "Collection Center"
partners with the 2 new added VAT accounts. Additionally, a new reconciliation
rules that detects the partner and the communication and creates a counter entry.
The VAT partner have been added to the VAT closing line for auto-reconciliation
with bank transaction.

task-5148985

closes odoo/enterprise#98224

Related: odoo/odoo#233360
Signed-off-by: John Laterre (jol) <jol@odoo.com>