[FIX] hr_contract_salary: fix styling of personal info in salary config
Before this commit, when website was not installed (or more precisely
html_builder), the salary configurator was not editable anymore due to
::before section that took the whole page. This was due to the changes made in
https://github.com/odoo/odoo/pull/229554/.
The style rule regarding relative positioned sections was move from html_editor
to html_builder as this was a more website specific (or html_builder specific)
rule.

The fix should then be done in modules that used that rules but that does not
depends on html_builder.

The commit applies relative position to the personal info section ans force
the ::before to let the pointer go through it to access underlying inputs.

Task-5154145

closes odoo/enterprise#96644

Signed-off-by: Jurgen Gjini (jugj) <jugj@odoo.com>