[FIX] web_gantt: change broken test workaround
Since a recent change with the way pointer targets are selected for DOM
events in tests, a workaround for a specific broken test in web_gantt
doesn't work anymore.

This commit changes the workaround by forcing a pill to be "interactive"
instead of ignoring its interactive status.

Note that the test case is still broken, and so is the actual scenario
it tries to replicate.

closes odoo/enterprise#97046

X-original-commit: c689af40607a51d2b2e6223ae691fae7d70a0487
Related: odoo/odoo#231313
Signed-off-by: Michaël Mattiello (mcm) <mcm@odoo.com>
Signed-off-by: Julien Mougenot (jum) <jum@odoo.com>