[FIX] stock_barcode_mrp*: improve barcode UI in dark mode for better visibility
*: stock_barcode

Before this commit:
=======================
When accessing a manufacturing order via Barcode in `dark mode`, the producing
bar of the finished product is difficult to read due to poor colour contrast.

Steps to reproduce:
=======================
1. Install the `mrp` module.
2. Switch to dark mode.
3. Open any manufacturing order using Barcode.
4. Observe the producing bar for the finished product.

- The text on the producing bar is hard to read because the colour of text does
not provide sufficient contrast.

- Additionally, for products tracked by `lot or serial number`,  when the same
product is added multiple times, the barcode view groups these entries into a
single grouped line, causing the same visibility issue in both MO and normal
`picking/transfer` operations.

Cause of the Issue:
=======================
In dark mode, both the text and background colour of the producing bar are
light, resulting in low colour contrast and making the text difficult to read.
Additionally, the `text-muted` class applied to the `units of measure (UoM)`
further reduces readability.

After this commit:
=======================
Updated the text colour of producing bar and background colour of grouped
lines to use a darker shade for better visibility in dark mode. Also, removed
the `text-muted` class used on `UOM` to improve redability.

TaskID-5117608

closes odoo/enterprise#96718

Signed-off-by: Quentin Wolfs (quwo) <quwo@odoo.com>