[FIX] account_online_synchronization: Set Journal company
Before this commit, when a user does a manual bank account
creation, we didn't pass the company id for the journal
creation. It implies that we use the environment value which
is problematic because it could be different from the one
account online link one leading to an error.
The aim of this commit is ensuring that we pass, as company_id
the same value as the one we have on account online link.

no task id

closes odoo/enterprise#96715

Signed-off-by: Ali Alfie (alal) <alal@odoo.com>