[FIX] pos_appointment: use total_capacity_reserved
The create actions in POS previously set a default total capacity
reserved (until [1]). But that field is not used in the form anywhere.

The field should be used when the capacity is being managed as it is
otherwise impossible to add or remove resources, since the booking lines
need to be adjusted based on the capacity reserved.

In cases where capacity is not managed, we can keep using
"waiting_list_capacity" as an indicator of how many people are expected
in a less strict setting.

Capacity reserved gets a default of 0 as it otherwise will not allow
quick-creating from the kanban and force open the form.

Wishlist gets its previous default of 2 which is assumed to be the most
common case.

The compute for the wishlist capacity is also updated so that when
the total capacity would be shown in the form instead, we always sync
the wishlist capacity based on it as then it makes no sense to allow
users to set an arbitrary value.

task-5103532

[1]: 2aab4dcfbe8491968f0721741efbab894667aefe

Part-of: odoo/enterprise#95539
Signed-off-by: Jérémy Hennecart (jeh) <jeh@odoo.com>