[FIX] pos_appointment: hide default button in form dialog
pos explicitly excludes the dialog.scss sheet from the web module. This
sheet however includes a rule to hide the default "ok" button when other
buttons are present on the dialog.

The rule is now copied in pos_appointment since that's where form
dialogs are introduced. It is not applied to other dialgos as pos makes
copious use of o-default-button even when they're not actually intended
to be hidden.

task-5103532

Part-of: odoo/enterprise#95539
Signed-off-by: Jérémy Hennecart (jeh) <jeh@odoo.com>