[FIX] hr_contract_salary: fix empty list when click on review button on employee
Issue:
When an offer is created and signed by the employee, the `Reviews` smart button
appears on the employee profile. Although it indicates that there are records,
Clicking the button redirects to an empty list.

Reason:
After converting contracts to the versioning system, the action for the Reviews
button does not properly display contract versions, regardless of whether they
are active or inactive, as it did in the previous version.

Fix:
Added a context active_test: False

task-5155455

closes odoo/enterprise#99782

X-original-commit: 22d66e5f74d2982425ff48e8549c60bc56267e87
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>
Signed-off-by: Priyanshu Goyal (prgo) <prgo@odoo.com>