[FIX] mrp_workorder: Set operator list to public employee
The task 4762527 added a warning for users that were not allowed to
access private employee profiles.
The "Add Operator" button in mrp_workorder (Shop Floor) was relying on
it.
I had to change their list model from hr.employee to hr.employee.public
This issues was raised by Aerospacelab in the following PR:
https://github.com/odoo/odoo/pull/228623

task-5156702

closes odoo/enterprise#97152

X-original-commit: 73f79021e7ec49028eaee77ac5849e87b71f2c04
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>