[FIX] spreadsheet_dashboard_crm: add operator, remove context
Some pivots in the CRM dashboards were missing the aggregator.
I manually added the new `sum_currency` aggregator because leads amounts
are not currency-converted. Note that this operator doesn't exists in the side
panel (for now).

This commit now filter the Pipeline dashboards on the "Last 12 months" instead
of since the begining of the universe (It's faster and it makes sense to see the
latest data).

This commit also remove some context keys that do not make sense:
- `search_default_assigned_to_me` => "My pipeline" filter would be applied when
  going to "See records"
- `default_team_id` => team with id '1' is a random team and might not even exist.

Task: 5158889
Part-of: odoo/enterprise#96810
Related: odoo/odoo#230918
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>