[FIX] documents: fix folders access when archiving a company
**Steps to reproduce:**
- Add some folders
- Create a new company
- Assign some folders to the new company
- Archive the company
- Click on Documents app
- AccessError is raised when opening it

**Issue:**
The issue seems to be related to caching issue on the field `type` when
fetching the documents with `search_panel_select_range` and going through
`_compute_display_name`:
`folders = accessible_records.filtered(lambda d: d.type == 'folder')`

This error was quite inconsistent and might be related to the cache
missing some prefetched data on a record it shouldn't have been
able to read. There is a need to ensure only the
folders available to the user are able to be fetched.

Filtering disabled companies is done by this logic:
`(company_id in self.env.companies or company_id not in self.env.user.company_ids)`
This allows documents from companies outside of the user own companies,
and only consider the user active companies for now.

**Fix:**
Tried to add the `type` field in the `search_panel_fields` but this might not
be needed if the issue comes from elsewhere.

Could also mitigate by explicitly checking for the user companies
in the domain of the searchs, but this needs to be checked as this behavior
might break other flows.

Adding the user archived companies to the other user companies will
remove this issue but we need to ensure this is the expected behavior.

Also need to ensure archived companies are not removed from the admin search
domain even if the user has `active_test=False` due to the security access domain:
`self_sudo = self.sudo().with_context(active_test=False)`

opw-4931278

closes odoo/enterprise#99126

X-original-commit: ef1afe9f0426dfc2ef58b0828e18896800ef2dea
Signed-off-by: Florian Charlier (flch) <flch@odoo.com>