[FIX] l10n_be_hr_payroll_acerta: Export Traceback
When generating the export file, a traceback occurred because
the name field was empty.
The logic has been updated to use the display_name field instead,
ensuring the export runs correctly, as it is always set to a value.
However, the display_name sometimes is just a date, so I depended
on a customizable new lable to show the employee to this version.

Additionally, a comprehensive test suite was added to validate
Acereta code behavior, including:
- validation of code length for employees, companies, and work entry types,
- handling of codes across different companies,
- and the complete Acereta export flow (from work entry creation
and validation to export file generation).

task-5153727

closes odoo/enterprise#100987

X-original-commit: f95378bc4358394333e32e70a9a3e7c9e00833c0
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>