[FIX] pos_settle_due: no error when commercial partner is not loaded
Before this commit, it was possible that commercial_partner_id was not
loaded, which caused an error when settling an invoice.

This commit fixes the issue by avoiding the need to load the full
partner record. Since only the partner ID is required to load the
account move, it is now read directly from the raw data, which already
includes the ID.

opw-5112883

closes odoo/enterprise#97828

X-original-commit: 33cc75426a4b08393d11443d6c70c23d34d7b31e
Signed-off-by: David Monnom (moda) <moda@odoo.com>
Signed-off-by: Pedram Bi Ria (pebr) <pebr@odoo.com>