[FIX] industry_fsm_report: ondelete method breaks uninstallation
Recently pull request https://github.com/odoo/enterprise/pull/86084
introduced an ondelete method on 'ir.model' that retrieves some
worksheet templates to delete them.  However, this method breaks when
uninstalling module 'worksheet':

    ir.model.data._module_data_uninstall():
        ... records are deleted ...
        ir.model.fields.unlink():
            drop column of corresponding fields
            delete ir.model.field records
        ir.model.unlink():
            drop table of corresponding models
            ir.model._unlink_if_uninstalling():
                self.env['worksheet.template'].search([('model_id', ...)]).unlink()
            delete ir.model records

The call to ir.model.unlink() crashes when searching for worksheet
templates, since column 'model_id' has been dropped already.  This makes
the transaction fail, and it is rolled back to a savepoint just before
the call to ir.model.unlink().  In other words, the uninstallation
manages to drop most of the columns that must go, but fails to drop all
the tables that must go.  And the uninstallation proceeds anyway...

Now consider uninstalling module 'resource'.  That module defines model
'resource.calendar' with required field 'name', and also defines a
record in that model (a default calendar).  When the module is
uninstalled, module 'worksheet' is also uninstalled (because it depends
on 'resource'), and so the situation above happens.  Consequently, most
of the columns of table 'resource_calendar' are dropped, but the table
is not.  If we reinstall module 'resource' after that, the ORM
re-creates column 'name' (which is NULL on the default calendar at
least), but fails to add the NOT NULL constraint on that column.

The fix consists in avoiding the search() above in the ondelete method
if the column 'model_id' does not exist anymore.

closes odoo/enterprise#97026

X-original-commit: d7f316951a144caef3d4cf7ba7dc9af07254741c
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>