[FIX] project_timesheet_forecast : Check if we are during working hours in planning reports
Steps to reproduce:
	- Create an employee with fixed working schedule from 8 to 5
	- Create a Planning shift for this employee that starts in a month and ends
	in the first day of the next month outside of working hours (e.g. Sept30th 8AM -> Oct1st 2AM)
	- Navigate to Timesheets / Planning analysis reports
	- Notice October has been taken into consideration in the report's planned hours

Cause:
The query we are using for the timesheets/planning report doesn't take
working hours into consideration it only cares about the date. So if the shift
ends in October 1st we are taking it into account whether it is inside
working hours or not.

Fix:
Add a condition to the where clause to check the working hours and if the
record lays in this period or not.

opw-5089052

closes odoo/enterprise#98601

X-original-commit: c5191732ba652637d9f56be5e7e374ff3534d9c3
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>