[FIX] account_accountant: partner_name
Before this commit, when a transaction had no partner_name and some lines with
the same partner. When unfolded, we had the info of the partner on the statement
line and on the line itself which was a duplicate of information.

This commit will change when the line is unfolded so that the partner is visible
on the statement line only when there is a partner_name

no task-id

Part-of: odoo/enterprise#98458
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>
Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>