[FIX] account_online_synchronization: Fixing duplicate banner
Before this commit, the action "Some transactions may be duplicates."
was broken because 2 elements from the old bank reconciliation widget
were removed. The commit [[1]] of the new bank reconciliation widget
forgot to add the action service and a link to the journal. These
2 elements were used to redirect the customer to the Find duplicates
tool.
This commit adds, in the extension of the bank reconciliation kanban
controller, the call to the action service and adapt the old code to
match the new one.

no task id

[1]: https://github.com/odoo/enterprise/commit/2335c953723dce66af8811fdfbfd5b811d42b109

closes odoo/enterprise#97110

X-original-commit: 61cce46b2347c114688de25789f5235b361ecbda
Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>