[FIX] pos_sale: Manage products paid on ecommerce in the POS
Because of https://github.com/odoo/odoo/pull/230809, it will cause the
quotations/orders menu to filter out any sale orders which have an
unpaid amount of 0.

The test_rental_pos test tries to create a sale order with a quantity of
0 and then settle it in the POS, but because of the filter, the sale
order will not show up in the menu causing the test to fail.

This PR is to increase the quantity in the sale order to 1 so that the
order shows up and the test can be executed properly.

Task-5138081

closes odoo/enterprise#96876

Related: odoo/odoo#230809
Signed-off-by: David Monnom (moda) <moda@odoo.com>