[FIX] voip: fix phone field margin when there is no flag
In the voip.call list view, when no flag is displayed, the phone number
is slightly offset to the right.

This is because the margin that separates the flag from the phone number
is still present.

This commit moves the margin from the phone number (always displayed) to
the flag (sometimes absent).

How to reproduce:
1. Make a call to an invalid phone number
2. Go to voip.call list view
3. Look at the numbers without a flag

closes odoo/enterprise#96877

Signed-off-by: Mohammed Basioni (basm) <basm@odoo.com>