[REV] hr_payroll: employees menu action
This reverts commit bbb22f4835197419be538a7e2a82c1743fac3fa3.

The fixes in the commit reverted is no longer needed thanks to the
recent changes by the framework to better manage the routing in backend.
Moreover, the fixes made by calling a client action to change the url
will trigger a hard reload and so the employees action will take more
time than before since all OWL components have to be mounted instead of
just mounting to needed owl components to load the Employees action.

closes odoo/enterprise#96967

Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>