[FIX] sale_planning: ensure unscheduled shift are immediately updated in gantt
Steps to reproduce:
Go to planning gantt view.
Click on shift.
Click on unschedule button on popover.

Issue:
When clicking on Unschedule button on a shift popover, sometimes the shift
still appears and again clicking on that it cause missing error occured until
the page is refreshed.

Cause:
The record wasn’t updating in the view after a shift was unscheduled, so the
unscheduled shift was still showing.

Fix:
Modified the method to reload the model and notify the view immediately
after calling action_unschedule, ensuring shifts reflect the latest server state
without requiring a manual refresh.

task-5075381

closes odoo/enterprise#102073

X-original-commit: 9ff0777f57927d2f8f82731856dc12f1242d6684
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>