[FIX] iap_extract: fix crash when clicking on another column of x2many
When a x2many field is filled using the rectangular selection tool, a
crash could occur when the component is busy filling the values for a
field and the user clicks on another column.

This happened because the `getNewRecordValues` function was using the
current value of `this.activeBoxType` to get the value for the new
record. This is incorrect as `this.activeBoxType` can change when
the user clicks on another column of the x2many.

task-none

Part-of: odoo/enterprise#96975
Signed-off-by: Ricardo Gomes Rodrigues (rigr) <rigr@odoo.com>