[PERF] account_report, l10n_be: remove no negative amount check
The "No negative amount in VAT report" return check was too slow,
taking about half of the total check refresh time. There is a now
a red banner on the report when there are negative amounts in it.

A method has been implemented to properly clean up return checks
that are no longer defined in code.

Task-id: 5145537
Part-of: odoo/enterprise#98633
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Jonathan Dauwe (jond) <jond@odoo.com>