[FIX] account_reports, l10n_be_reports: Branch refresh checks
Users needed access rights to all child companies to refresh
the checks.

On the other hand, users who are not administrators and
have only access to the child companies were not
able to access the dashboard (if there are returns in the
root company) and the return view.

Users can only submit a return if they have access to the
parent company and all its branches.
This prevents submissions without reviewing data from all
related companies.

Task-id: 5145537
Part-of: odoo/enterprise#98633
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Jonathan Dauwe (jond) <jond@odoo.com>