[FIX] web_studio: normalize: fix base_arch getter
Since commit 52f27c457e50a61c3b11c59e3fdd5ddd383f1317, the normalize function
applies on `self` (ir.ui.view) and can take a string representing the arch to normalize.

The normalization then occurs on the combination of the two, outputing a simplified version
of the arch passed as a string in arguments.

The base arch (`combined_arch`) is retrieved from self.
Before this commit, this was done by controlling on self.inherit_id, which did not cover
all bases to retrieve the main view's combined arch without the studio inheritance.
After this commit, this is done by controlling on self.mode, because an inherited view can also
be a root view.

opw-5154217

closes odoo/enterprise#96992

Signed-off-by: Bastien Pierre (ipb) <ipb@odoo.com>